# api/mtgcard.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mtgcard.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/inventory": {
      "post": {
        "operationId": "DeleteInventory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiInventoryResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apideleteInventoryParams"
            }
          }
        ],
        "tags": [
          "MTGCardService"
        ]
      }
    },
    "/v1/inventory/{inventory_id}": {
      "get": {
        "operationId": "GetMTGCard",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiMTGCardResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "inventory_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "MTGCardService"
        ]
      },
      "post": {
        "operationId": "AddMTGCard",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiMTGCardResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "inventory_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiaddMTGCardParams"
            }
          }
        ],
        "tags": [
          "MTGCardService"
        ]
      }
    },
    "/v1/inventory/{inventory_id}/{name}": {
      "delete": {
        "operationId": "DeleteMTGCard",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apideleteResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "inventory_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "qty",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MTGCardService"
        ]
      }
    }
  },
  "definitions": {
    "apiInventoryResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "apiMTGCard": {
      "type": "object",
      "properties": {
        "inventory_id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "qty": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apiMTGCardResponse": {
      "type": "object",
      "properties": {
        "cards": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiMTGCard"
          }
        }
      }
    },
    "apiaddInventoryParams": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "apiaddMTGCardParams": {
      "type": "object",
      "properties": {
        "inventory_id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "qty": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apideleteInventoryParams": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "apideleteResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "apigetInventoryParams": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}